requirements.txt
requests
feedparser
pandas
python-dateutil